when panning tide boxes and current arrows aren't correctly display in opengl mode with accelerated panning enable.
In this case the viewpoint for the redrawn area can be small
with current/tide point outside.

I think the worst case is Raz Blanchard (49.75N 2W) up to 12 knots
current heading north, that's a 0.2 degree arrow.

There's still an issue though, panning changes the chart scale and
around the limit between full arrows and squares previously current
drawn as an arrow are redrawn as a square and the arrow is half drawn.